Fix cast warning when calculating camera arm length for Ortho Near/Far plane correction.
#rb Carl.Lloyd

[CL 31287561 by jon cain in ue5-main branch]

+++ Engine/Source/Runtime/Engine/Classes/Camera/CameraTypes.h
@@ -166,7 +166,7 @@ public:
 	/** Sets the camera distance from view target for AutoCalculateOrthoPlanes */
 	ENGINE_API inline void SetOrthoCameraArmLengthFromOwnerLocation(const FVector ActorLocation)
 	{
-		OrthoCameraArmLength = (ActorLocation - Location).Length();
+		OrthoCameraArmLength = static_cast<float>((ActorLocation - Location).Length());
 	}
 };
